Add files via upload
Project 1 - Statistical Learning Project 
PROJECT OBJECTIVE: Company’s management wants to invest on proposals on managing some of the best teams in the league. The analytics 
department has been assigned with a task of creating a report on the performance shown by the teams. Some of the older teams are already in 
contract with competitors. Hence Company X wants to understand which teams they can approach which will be a deal win for them