Add files via upload
CONTEXT : Medical research university X is undergoing a deep research on patients with certain conditions. University has an internal AI team. Due to confidentiality the patient’s details and the conditions are masked by the client by providing different datasets to the AI team for developing a AIML model which can predict the condition of the patient depending on the received test results.

DATA DESCRIPTION : The data consists of biomechanics features of the patients according to their current conditions. Each patient is represented in the data set by six biomechanics attributes derived from the shape and orientation of the condition to their body part.

PROJECT OBJECTIVE :To Demonstrate the ability to fetch, process and leverage data to generate useful predictions by training Supervised Learning algorithms.